INTERLOK-1083 Rename from Codec to Transcode
