+++ adapter/src/test/java/com/adaptris/core/services/transcode/TranscodingServiceCase.java
@@ -1,17 +1,17 @@
-package com.adaptris.core.services.codec;
+package com.adaptris.core.services.transcode;
 
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.AdaptrisMessageFactory;
 import com.adaptris.core.MimeEncoder;
 import com.adaptris.core.ServiceCase;
 
-public abstract class CodecServiceCase extends ServiceCase {
+public abstract class TranscodingServiceCase extends ServiceCase {
 
   /**
    * Key in unit-test.properties that defines where example goes unless overriden {@link #setBaseDir(String)}.
    *
    */
-  public static final String BASE_DIR_KEY = "CodecServiceExamples.baseDir";
+  public static final String BASE_DIR_KEY = "TranscodingServiceExamples.baseDir";
 
   static final String TEST_METADATA_KEY = "helloMetadataKey";
   static final String TEST_METADATA_KEY_2 = "worldMetadataKey";
@@ -19,7 +19,7 @@ public abstract class CodecServiceCase extends ServiceCase {
   static final String TEST_METADATA_VALUE_2 = "World";
   static final String TEST_PAYLOAD = "The quick brown fox jumped over the lazy dog.";
 
-  public CodecServiceCase(String name) {
+  public TranscodingServiceCase(String name) {
     super(name);
     if (PROPERTIES.getProperty(BASE_DIR_KEY) != null) {
       setBaseDir(PROPERTIES.getProperty(BASE_DIR_KEY));
