+++ adapter/src/test/resources/unit-tests.properties.template
@@ -88,7 +88,7 @@ junit.CacheServiceExamples.baseDir=${junit.ServiceCase.baseDir}/cache
 junit.AggregatingServiceExamples.baseDir=${junit.ServiceCase.baseDir}/aggregator
 junit.XmlServiceExamples.baseDir=${junit.ServiceCase.baseDir}/xml
 junit.RegexpServiceExamples.baseDir=${junit.ServiceCase.baseDir}/regexp
-junit.CodecServiceExamples.baseDir=${junit.ServiceCase.baseDir}/codec
+junit.TranscodingServiceExamples.baseDir=${junit.ServiceCase.baseDir}/transcoding
